+++ pyo3-derive-backend/src/pyclass.rs
@@ -273,8 +273,6 @@ fn impl_class(
                             if FREELIST.is_null() {
                                 FREELIST = Box::into_raw(Box::new(
                                     pyo3::freelist::FreeList::with_capacity(#freelist)));
-
-                                <#cls as pyo3::type_object::PyTypeObject>::init_type();
                             }
                             &mut *FREELIST
                         }
@@ -372,7 +370,7 @@ fn impl_class(
     };
 
     Ok(quote! {
-        impl pyo3::type_object::PyTypeInfo for #cls {
+        unsafe impl pyo3::type_object::PyTypeInfo for #cls {
             type Type = #cls;
             type BaseType = #base;
             type ConcreteLayout = pyo3::pyclass::PyClassShell<Self>;
@@ -384,10 +382,23 @@ fn impl_class(
             const FLAGS: usize = #(#flags)|* | #extended;
 
             #[inline]
-            fn type_object() -> *mut pyo3::ffi::PyTypeObject {
-                static TYPE_OBJECT: pyo3::derive_utils::LazyTypeObject =
-                    pyo3::derive_utils::LazyTypeObject::new();
-                TYPE_OBJECT.get()
+            fn type_object() -> std::ptr::NonNull<pyo3::ffi::PyTypeObject> {
+                use std::ptr::NonNull;
+                use pyo3::type_object::LazyTypeObject;
+                static TYPE_OBJECT: LazyTypeObject = LazyTypeObject::new();
+                TYPE_OBJECT.get_or_init(|| {
+                        // automatically initialize the class on-demand
+                        let gil = pyo3::Python::acquire_gil();
+                        let py = gil.python();
+                        let boxed = pyo3::pyclass::create_type_object::<Self>(py, Self::MODULE)?;
+                        Ok(unsafe { NonNull::new_unchecked(Box::into_raw(boxed)) })
+                    })
+                    .unwrap_or_else(|e| {
+                        let gil = Python::acquire_gil();
+                        let py = gil.python();
+                        e.print(py);
+                        panic!("An error occurred while initializing class {}", Self::NAME)
+                    })
             }
         }